requirements.txt
quart
requests
Pillow
